I will implement the Context Engine plugin architecture with lifecycle hooks. Here is my approach:
- Refactor `magda_agent/memory/context_engine.py` to support `bootstrap`, `ingest`, `assemble`, `compact`, `before_retrieval`, `after_retrieval`, and `on_context_update` hooks.
- Move `DefaultContextPlugin` to `magda_agent/memory/default_context_plugin.py` and update its implementation.
- Update `Consciousness` and `WorkingMemory` call-sites to use the new hooks.
- Consolidate imports and remove the redundant `magda_agent/context/` directory.
- Update tests in `tests/test_context_engine.py` and `tests/test_continuity.py`.
- Mark task `context-engine-plugin-impl` as finished in `agent_tasks.json`.
