+++ magda_agent/memory/context_engine.py
@@ -1,8 +1,24 @@
 import logging
-from typing import List, Protocol, Any, Callable
+from typing import List, Protocol, Any, Callable, Dict, Optional
 
 class ContextPlugin(Protocol):
     """Protocol defining the lifecycle hooks for a Context Engine plugin."""
+    async def bootstrap(self, config: Dict[str, Any]) -> None:
+        """Initialize the plugin with configuration."""
+        ...
+
+    async def ingest(self, content: str, metadata: Dict[str, Any]) -> str:
+        """Process incoming content before it is stored or used."""
+        ...
+
+    async def assemble(self, context_items: List[Any], metadata: Dict[str, Any]) -> str:
+        """Assemble the context string from retrieved items for the LLM."""
+        ...
+
+    async def compact(self, context_items: List[Any], metadata: Dict[str, Any]) -> List[Any]:
+        """Compact or summarize the context when limits are reached."""
+        ...
+
     def before_retrieval(self, query: str, user_id: int) -> str:
         """Called before context is retrieved. Can modify the query."""
         ...
@@ -21,13 +37,51 @@ class ContextEngine:
     ContextEngine manages context dynamically using a plugin architecture
     with lifecycle hooks.
     """
-    def __init__(self) -> None:
-        self._plugins: List[ContextPlugin] = []
+    def __init__(self, plugins: Optional[List[ContextPlugin]] = None) -> None:
+        self._plugins: List[ContextPlugin] = plugins or []
 
     def register_plugin(self, plugin: ContextPlugin) -> None:
         """Registers a new plugin with the context engine."""
         self._plugins.append(plugin)
         logging.debug(f"Registered plugin: {plugin.__class__.__name__}")
+
+    def add_plugin(self, plugin: ContextPlugin) -> None:
+        """Alias for register_plugin for compatibility."""
+        self.register_plugin(plugin)
+
+    async def bootstrap_all(self, config: Dict[str, Any]) -> None:
+        """Initialize all plugins."""
+        for plugin in self._plugins:
+            if hasattr(plugin, 'bootstrap'):
+                await plugin.bootstrap(config)
+
+    async def ingest(self, content: str, metadata: Dict[str, Any]) -> str:
+        """Run content through ingest hook of all plugins."""
+        current_content = content
+        for plugin in self._plugins:
+            if hasattr(plugin, 'ingest'):
+                current_content = await plugin.ingest(current_content, metadata)
+        return current_content
+
+    async def assemble(self, context_items: List[Any], metadata: Dict[str, Any]) -> str:
+        """Assemble context using all plugins."""
+        assembled_context = ""
+        # If no plugins, default behavior
+        if not self._plugins:
+            return "\n".join([str(item) for item in context_items])
+
+        for plugin in self._plugins:
+            if hasattr(plugin, 'assemble'):
+                assembled_context = await plugin.assemble(context_items, metadata)
+        return assembled_context
+
+    async def compact(self, context_items: List[Any], metadata: Dict[str, Any]) -> List[Any]:
+        """Compact context through all plugins."""
+        current_items = context_items
+        for plugin in self._plugins:
+            if hasattr(plugin, 'compact'):
+                current_items = await plugin.compact(current_items, metadata)
+        return current_items
 
     def retrieve_context(self, query: str, user_id: int, base_retrieval_func: Callable[[str, int], List[Any]]) -> List[Any]:
         """
@@ -36,25 +90,19 @@ class ContextEngine:
         """
         current_query = query
         for plugin in self._plugins:
-            current_query = plugin.before_retrieval(current_query, user_id)
+            if hasattr(plugin, 'before_retrieval'):
+                current_query = plugin.before_retrieval(current_query, user_id)
 
         context = base_retrieval_func(current_query, user_id)
 
         for plugin in self._plugins:
-            context = plugin.after_retrieval(context, current_query, user_id)
+            if hasattr(plugin, 'after_retrieval'):
+                context = plugin.after_retrieval(context, current_query, user_id)
 
         return context
 
     def update_context(self, new_context: Any, user_id: int) -> None:
         """Triggers the on_context_update hook for all registered plugins."""
         for plugin in self._plugins:
-            plugin.on_context_update(new_context, user_id)
-
-    async def compact(self, entries: List[Any], context_params: dict) -> List[Any]:
-        """Compacts entries when memory limit is reached (used by working memory)."""
-        limit = context_params.get("limit", len(entries))
-        # Default behavior: remove the oldest entry to satisfy limit
-        # Plugins can override this behavior in the future via hooks.
-        if len(entries) > limit:
-            return entries[1:]
-        return entries
+            if hasattr(plugin, 'on_context_update'):
+                plugin.on_context_update(new_context, user_id)
